backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
